requirements.txt
python-dotenv
requests
pandas
pytz
psycopg2
ratelimit
tqdm